backend/data/data_fetcher.py
# Placeholder for data fetching functions (e.g., from FRED, Yahoo Finance, Shiller dataset)
import pandas as pd
import requests
import os
from dotenv import load_dotenv

load_dotenv() # Load environment variables if using API keys

FRED_API_KEY = os.getenv("FRED_API_KEY")
ALPHA_VANTAGE_API_KEY = os.getenv("ALPHA_VANTAGE_API_KEY")

def fetch_shiller_data():
    """Fetches and preprocesses Shiller's U.S. Stock Market Data."""
    # Logic to download and parse the Shiller dataset (e.g., from an online source)
    # Example: url = "http://www.econ.yale.edu/~shiller/data/ie_data.xls"
    # df = pd.read_excel(url, sheet_name='Data', header=7)
    # ... preprocessing steps ...
    print("Fetching Shiller data (placeholder)")
    # Return a pandas DataFrame
    return pd.DataFrame() # Placeholder

def fetch_fred_data(series_id):
    """Fetches data for a specific series ID from FRED."""
    if not FRED_API_KEY:
        print("FRED API Key not found. Set FRED_API_KEY environment variable.")
        return pd.DataFrame()
    
    url = f"https://api.stlouisfed.org/fred/series/observations?series_id={series_id}&api_key={FRED_API_KEY}&file_type=json"
    try:
        response = requests.get(url)
        response.raise_for_status() # Raise an exception for bad status codes
        data = response.json()
        # Convert to DataFrame
        df = pd.DataFrame(data['observations'])
        df = df[['date', 'value']]
        df['date'] = pd.to_datetime(df['date'])
        df['value'] = pd.to_numeric(df['value'], errors='coerce') # Handle non-numeric if necessary
        df = df.set_index('date')
        print(f"Fetched {series_id} data from FRED")
        return df
    except requests.exceptions.RequestException as e:
        print(f"Error fetching {series_id} from FRED: {e}")
        return pd.DataFrame()
    except Exception as e:
        print(f"Error processing FRED data for {series_id}: {e}")
        return pd.DataFrame()

def fetch_sp500_data():
    """Fetches S&P 500 data (e.g., using FRED or Alpha Vantage)."""
    # Example using FRED's SP500 index
    return fetch_fred_data("SP500")

def fetch_cpi_data():
    """Fetches CPI data (e.g., using FRED's CPIAUCSL)."""
    return fetch_fred_data("CPIAUCSL")

# Add functions for other data sources (GDP, interest rates, etc.) as needed

def fetch_ohlc_data(ticker: str, start_date: str, end_date: str):
    """Fetches OHLCV data for a given ticker using yfinance."""
    try:
        import yfinance as yf
        stock = yf.Ticker(ticker)
        # Add 1 day to end_date because yfinance excludes the end date
        end_date_yf = (pd.to_datetime(end_date) + pd.Timedelta(days=1)).strftime('%Y-%m-%d')
        hist = stock.history(start=start_date, end=end_date_yf)
        if hist.empty:
            print(f"No data found for {ticker} between {start_date} and {end_date}")
            return pd.DataFrame()
        # Select and rename columns for consistency if needed
        hist = hist[['Open', 'High', 'Low', 'Close', 'Volume']]
        # Convert index to string date format if required by frontend
        # hist.index = hist.index.strftime('%Y-%m-%d')
        print(f"Fetched OHLCV data for {ticker} from yfinance")
        return hist
    except ImportError:
        print("Error: yfinance library not installed. Run 'pip install yfinance'")
        return pd.DataFrame()
    except Exception as e:
        print(f"Error fetching OHLCV data for {ticker}: {e}")
        return pd.DataFrame()

if __name__ == '__main__':
    # Example usage for OHLC
    ohlc_df = fetch_ohlc_data("SPY", "2023-01-01", "2023-01-10") # Example ticker SPY
    print("\nOHLC Data (SPY Example):")
    print(ohlc_df.head())

    # Example usage
    sp500_df = fetch_sp500_data()
    print("\nS&P 500 Data:")
    print(sp500_df.tail())

    cpi_df = fetch_cpi_data()
    print("\nCPI Data:")
    print(cpi_df.tail())

    # shiller_df = fetch_shiller_data()
    # print("\nShiller Data (Placeholder):")
    # print(shiller_df.head())